+++ src/generic_esp_32/generic_esp_32.cpp
@@ -24,8 +24,8 @@
 
 extern "C"
 {
-#include "../specific_m5coreink/powerpin.h"
-#include "../specific_m5coreink/rtc.h"
+#include <specific_m5coreink/powerpin.h>
+#include <specific_m5coreink/rtc.h>
 }
 
 #ifdef CONFIG_TWOMES_PROV_TRANSPORT_BLE
